services/recipe-service/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.2
SQLAlchemy==2.0.19
psycopg2-binary==2.9.7
python-dotenv==1.0.0
numpy==1.24.4
pandas==2.0.3
torch==2.0.1
transformers==4.30.2
qdrant-client==1.4.0
sentence-transformers==2.2.2